\name{BRCA.100}
\alias{BRCA.100}
\docType{data}
\title{
Breast Cancer gene expression data.
}
\description{
Expression data of 100 genes (rows) in 533 conditions (columns).
}
\usage{data("BRCA.100")}
\format{
  A data frame with 533 observations on 100 variables.
}
\source{
Normalized microarray data from TCGA - https://tcga-data.nci.nih.gov/tcga/.
}
\examples{
data(BRCA.100)
}
\keyword{datasets}
