\name{BRCA.results}
\alias{BRCA.results}
\docType{data}
\title{
Breast Cancer \code{\link{clusterRange}} output
}
\description{
Output of \code{\link{clusterRange}} on breast cancer gene expression dataset.
To be displayed with \code{\link{plotRange3D}} for visualization.
}
\usage{data("BRCA.results")}
\format{
  List of 2. First element is a list of 11 \code{\linkS4class{COMMUNAL}} objects. Second element is a vector of the corresponding number of data points that were clustered.
}
\examples{
data(BRCA.results)
}
\keyword{datasets}
